+++ rnavigate/analysis/__init__.py
@@ -1,5 +1,6 @@
 from .logcompare import LogCompare
 from .lowss import LowSS
 from .deltashape import DeltaSHAPE
+from .auroc import WindowedAUROC
 
-__all__ = ["LogCompare", "LowSS", "DeltaSHAPE"]
+__all__ = ["LogCompare", "LowSS", "DeltaSHAPE", "WindowedAUROC"]