+++ docs/changelog.md
@@ -11,4 +11,5 @@ CHANGES:
 - removed py3dmol version from env.yaml (uses most recent)
 - Added ROC plots
 - Added DeltaSHAPE
+- Added Windowed AUROC